refactor: ConnectionType 디렉토리 변경
