+++ src/main/java/trend_setter/turtlerun/connection/entity/Connection.java
@@ -17,6 +17,7 @@ import java.util.List;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import trend_setter.turtlerun.connection.constant.ConnectionType;
 import trend_setter.turtlerun.global.common.BaseEntity;
 import trend_setter.turtlerun.user.entity.User;
